requirements.txt
python>=3.9
pytorch>=1.10.0
torchvision>=0.11.0

joblib
scikit-image
matplotlib
opencv-python
tqdm
tensorflow
pyyaml
tensorboardx
wandb
pycocotools
classy_vision

# FastSAM-------------------------------------

matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0

pandas>=1.1.4
seaborn>=0.11.0

gradio==3.35.2

# Ultralytics-----------------------------------
ultralytics == 8.0.120